Gate #[MapRequestPayload(type: ...)] test on symfony/http-kernel 7.1
The $type argument of #[MapRequestPayload] only exists since
symfony/http-kernel 7.1. On older versions the attribute has no such
constructor parameter, so no usage is emitted (correctly matching the
runtime, where array payload mapping does not exist) and the shared
fixture failed on PHP 8.1 (http-kernel downgraded to 6.4) and
prefer-lowest jobs. Move the test case into a version-gated fixture.
